requirements.txt
# Google GenAI SDK
google-genai

# Model Context Protocol (MCP)
mcp
pydantic
python-dotenv

# Cisco pyATS & Genie
pyats[full]

# Async support
asyncio